requirements.txt
Django>=1.4,<1.5
MySQL-python
South==0.7.6
django-mailchimp-v1.3==1.3.0
lyrical_page==2.0.0p3
django-grappelli==2.4.4
uwsgi
